+++ .gitignore
@@ -3,6 +3,7 @@ node_modules
 /build
 /.svelte-kit
 .svelte-kit
+/.svelte-kit*
 .svelte-kit*
 /package
 .env
