+++ .svelte-kit/types/src/routes/api/$types.d.ts
@@ -5,6 +5,15 @@ type Expand<T> = T extends infer O ? { [K in keyof O]: O[K] } : never;
 type MatcherParam<M> = M extends (param : string) => param is infer U ? U extends string ? U : string : string;
 type RouteParams = {  };
 type RouteId = '/api';
+type MaybeWithVoid<T> = {} extends T ? T | void : T;
+export type RequiredKeys<T> = { [K in keyof T]-?: {} extends { [P in K]: T[K] } ? never : K; }[keyof T];
+type OutputDataShape<T> = MaybeWithVoid<Omit<App.PageData, RequiredKeys<T>> & Partial<Pick<App.PageData, keyof T & keyof App.PageData>> & Record<string, any>>
+type EnsureDefined<T> = T extends null | undefined ? {} : T;
+type OptionalUnion<U extends Record<string, any>, A extends keyof U = U extends U ? keyof U : never> = U extends unknown ? { [P in Exclude<A, keyof U>]?: never } & U : never;
+export type Snapshot<T = any> = Kit.Snapshot<T>;
+type PageParentData = EnsureDefined<import('../$types.js').LayoutData>;
 
+export type PageServerData = null;
+export type PageData = Expand<PageParentData>;
 export type RequestHandler = Kit.RequestHandler<RouteParams, RouteId>;
 export type RequestEvent = Kit.RequestEvent<RouteParams, RouteId>;
