+++ src/config/settings.js
@@ -1,0 +1,13 @@
+
+const firebaseConfig = {
+  apiKey: "AIzaSyC7ij7x2up4mqxJvnArSI-adaDxPhUzt0k",
+  authDomain: "gns-fub-chat-45501.firebaseapp.com",
+  databaseURL: "https://gns-fub-chat-45501-default-rtdb.firebaseio.com",
+  projectId: "gns-fub-chat-45501",
+  storageBucket: "gns-fub-chat-45501.appspot.com",
+  messagingSenderId: "502025304899",
+  appId: "1:502025304899:web:2fb48774eb1cf125744df6",
+  measurementId: "G-19HXS5CKH2"
+};
+
+export { firebaseConfig }
